Added GAL unit in Unit enum class

+++ src/com/tw/step8/assignment3/Length.java
@@ -14,11 +14,11 @@ public class Length {
   }
 
   public int compare(Length otherLength) {
-    return Integer.compare(this.toMM(), otherLength.toMM());
+    return Double.compare(this.toMM(), otherLength.toMM());
 
   }
 
-  public int toMM() {
+  public double toMM() {
     return this.superMeasure.toMM() + this.subMeasure.toMM();
   }
 }
